+++ adles/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = "1.4.4"
+__version__ = "1.4.5"
 __author__ = "Christopher Goes"
 __email__ = "goesc@acm.org"
 __license__ = "Apache 2.0"