+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [1.4.5] - 2018-03-17
+### Added (UNTESTED)
+- Progress bars to: `clone-vms`, `vm-power`, `vm-snapshots`
+- Dependency on [tqdm](https://github.com/tqdm/tqdm) to provide terminal
+progress bars for long-running operations (such as cloning VMs).
+
+
 ## [1.4.4] - 2018-03-11
 ### Added
 - New argument: `--syslog`. Configures saving of log output to the specified Syslog server.